# Optimize putVarInt