Some evidence of the redundancy of `.&. 127`:
```
> quickCheck (\(x :: Word64) -> (fromIntegral (x .|. 128) :: Word8) == (fromIntegral (x .&. 127 .|. 128) :: Word8))
+++ OK, passed 100 tests.
```