@ethercrow 2020-05-27
@blackgnezdo I failed to build the project at all. Both cabal and stack complain about `unexpected ","` in proto-lens-runtime.cabal