@ethercrow 2020-05-27
Another `putVarInt` idea is:

```
putVarInt n | n < 128 = word8 (fromIntegral n)
putVarInt n = putVarIntRecursive n
{-# INLINE putVarInt #-}

putVarIntRecursive = ... current implementation ...
```

So that GHC can optimize code like `putVarInt <some-small-constant>`, I have lots of it in https://github.com/ethercrow/lightstep-haskell/blob/master/gen/Proto/Collector.hs

But again it's pure speculation that it doesn't. I'm only assuming the lack of inlining because the function is recursive.