@ethercrow 2020-05-27
I looked at godbolt.org with `-O2`, it shows the difference of

```
        orq $128,%rax
```

vs

```
        andl $127,%eax
        orq $128,%rax
```