@judah 2020-05-28
Thanks @ethercrow for the patch! This change looks reasonable to me. One small request: can you please add a comment indicating that fromIntegral does the "and" operation for us? I think that would make it a little easier to compare to other implementations of the encoding.