@ethercrow 2020-05-28
@judah Sure, I added a comment.

I also went ahead with the inlining change. Couldn't get tests to run locally, but hopefully Travis will tell me if everything's fine.