@judah 2020-05-28
@ethercrow, for your second proposed change, I think we'll want to run more detailed analysis (actual benchmarking and looking at Core) to confirm that it's a net improvement. I'm also wondering whether `INLINABLE` would be more appropriate than `INLINE`. Maybe we should split it out into a second PR?

To benchmark the change you ought to be able to run `stack bench proto-lens-benchmarks --ghc-options=-O2`. (You'll also need to run the Setup instructions here: https://github.com/google/proto-lens#setup) Generally, running the benchmark before/after a change and comparing the results can give some useful information (although it's sometimes a little noisy due to GC, local machine fluctuations, etc.)

If that doesn't work for you, and you're interested in continuing to work on it, please let us know the exact error message you're getting and we can help diagnose it further.