@ethercrow 2020-05-28
I got `stack test` and `stack bench` working, here are the encoding numbers:

This PR (but with INLINABLE) | This PR (with INLINE) | Master

<img width="1533" alt="Screen Shot 2020-05-28 at 9 59 18 PM" src="https://user-images.githubusercontent.com/222467/83187531-98106400-a12e-11ea-9de1-32c6d86c9080.png">

Apply grains of salt as needed.