@judah 2020-05-29
Thanks for those benchmark results! It does look like we get some significant improvements on int32-packed,  (~5.8us-> ~2.0us), no-unused (57us -> 21us) and with-unused (60us -> 21us), which is expected since their encodings are particularly tag-heavy. And it seems INLINE is necessary rather than INLINABLE.

I did see int32-packed get *slower* in speed, which is a bit unexpected since we'd expect the bulk of the time to be copying the packed bytes. Possibly that was just noise though.

Once concern I have is that the INLINE causes more code to be generated and compilation time go up (for example, when varint-encoding an arbitrary int). As an experiment I hacked proto-lens-protoc to only build Proto.Google.Protobuf.Descriptor, and saw that compiling it was consistently ~16% slower: 38s at HEAD vs 44s with this change. I'm guessing this is due to the extra inlining making the generated code noticeably larger.

@blackgnezdo , any thoughts?