@blackgnezdo 2020-05-29
Does compile time change uniformly across GHC versions? Is there a known GHC issue which could address slow compilation?

I don't want to make a trade-off between compilation speed and runtime speed for the users. I also don't like extra knobs and this one will probably be on the uglier side (CPP?).

One possible data point would be to try on the change on a bigger sample of protos. I'll try to find a willing internal volunteer to see what the impact is for us.