@ethercrow 2020-05-29
Just my unsolicited opinion: if compilation speed is important to your organization, please contribute some developer time or sponsor some existing devs who optimize GHC in their free time. Seeing issues like https://gitlab.haskell.org/ghc/ghc/-/merge_requests/3363 I wonder how many more of such small things are waiting there just because so few people are looking for them.