@ethercrow 2020-08-11
Weekly bump.