@judah 2020-08-14
@ethercrow I apologize for the delay in reply.  Unfortunately I personally won't have much bandwidth to investigate the compilation time slowdown in the near term. Also, as I'm not at Google anymore I don't have access to their internal protos to do more thorough testing.

I think your first change (removing `.&. 127`) is uncontroversial; how about we merge that one for now? I opened #396 for that.