@judah 2020-08-14
Incidentally, here's one idea that I had, but didn't have time to try out. Our use of `putVarInt` with constants is all in *generated* code (output from `proto-lens-protoc`). See for example:
https://github.com/google/proto-lens/blob/0ac7a36e91038ce40a9e4108f36e75960281224e/proto-lens-protoc/app/Data/ProtoLens/Compiler/Generate/Encoding.hs#L556

So instead of relying on the INLINE mechanism, we could make your proposed transformation entirely within the code generation stage. More precisely, everywhere that we currently use `putVarInt' @@ int ...` (which generates the Haskell code `putVarInt {n}`) we could replace with a function that special-cases it to generate `putWord8 {n}`. For example, somethng like:

```
putVarIntConst :: Integer -> HsExpr'
putVarIntConst n
  | n < 128 = var "putWord8" @@ int n
  | otherwise = var "putVarInt" @@ int n
```

That should have essentially no impact on compilation times, and compile down to the same code as your proposed patch.