@ethercrow 2020-08-14
> So instead of relying on the INLINE mechanism, we could make your proposed transformation entirely within the code generation stage.

👍